:recycle: [Refactor]: FSD 구조 적용 - 4. shared 레이어 unpack 마무리
