+++ apps/client/src/pages/Broadcast/config/index.ts
@@ -1,0 +1,1 @@
+export { ENCODING_OPTIONS, RESOLUTION_OPTIONS } from './videoOptions';
