+++ apps/client/src/pages/Broadcast/types/trackTypes.ts
@@ -1,0 +1,5 @@
+export type Tracks = {
+  video: MediaStreamTrack | undefined;
+  mediaAudio: MediaStreamTrack | undefined;
+  screenAudio: MediaStreamTrack | undefined;
+};
