+++ apps/client/src/shared/lib/utils/utils.ts
@@ -11,21 +11,3 @@ export const checkDependencies = (functionName: string, dependencies: { [key: st
   if (missing.length === 0) return null;
   return new Error(`${functionName} Error: ${missing.join(',')}이(가) 없습니다.`);
 };
-
-export const getPayloadFromJWT = () => {
-  const token = localStorage.getItem('accessToken');
-  if (!token) return undefined;
-  const base64Payload = token.split('.')[1];
-  const base64 = base64Payload.replace(/-/g, '+').replace(/_/g, '/');
-
-  const decodedJWT = JSON.parse(
-    decodeURIComponent(
-      window
-        .atob(base64)
-        .split('')
-        .map(c => `%${  (`00${  c.charCodeAt(0).toString(16)}`).slice(-2)}`)
-        .join(''),
-    ),
-  );
-  return decodedJWT;
-};
