+++ apps/client/src/shared/types/sharedTypes.ts
@@ -1,0 +1,1 @@
+export type Field = 'WEB' | 'AND' | 'IOS' | '';
